backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=25.1.0",
    "aiosqlite>=0.22.1",
    "anthropic>=0.79.0",
    "fastapi>=0.128.5",
    "google-generativeai>=0.8.6",
    "httpx>=0.28.1",
    "mistralai>=1.12.0",
    "openai>=2.17.0",
    "pillow>=12.1.0",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.46",
    "sse-starlette>=3.2.0",
    "uvicorn[standard]>=0.40.0",
]
